backend/pyproject.toml
[tool.poetry]
name = "photoshare-backend"
version = "0.1.0"
description = "Backend server for PhotoShare application"
authors = ["Your Name <your.email@example.com>"]
package-mode = false
packages = [{include = "app"}]

[tool.poetry.scripts]
server = "app.main:start_server"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.25"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
boto3 = "^1.36.26"
bcrypt = "^4.2.1"
python-jose = "^3.4.0"
cryptography = "^44.0.1"
pyjwt = "^2.10.1"
passlib = "^1.7.4"
jwt = "^1.3.1"

[tool.poetry.group.dev.dependencies]
black = "^24.1.0"
flake8 = "^7.0.0"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"